Fix: Segurança nas ações de redes

+++ app/Http/Controllers/NetworkController.php
@@ -7,6 +7,7 @@ use App\Repositories\NetworkRepository;
 use App\Rules\NetworkRule;
 use App\Services\NetworkService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 
@@ -28,7 +29,7 @@ class NetworkController
     public function index(Request $request)
     {
         try {
-            $networks = $this->repository->getAllByEnterprise($request->user()->enterprise_id, ['member', 'office']);
+            $networks = $this->repository->getAllByEnterprise($request->user()->enterprise_id, ['member', 'congregation']);
             $filledData = EnterpriseHelper::filledData($request->user()->enterprise_id);
 
             return response()->json(['filled_data' => $filledData, 'networks' => $networks], 200);
@@ -43,12 +44,16 @@ class NetworkController
     {
         try {
             DB::beginTransaction();
+            if (Auth::user()->enterprise && Auth::user()->enterprise->created_by !== null) {
+                throw new \Exception('Sua organização não tem permissão');
+            }
+
             $network = $this->service->create($request);
 
             if ($network) {
                 DB::commit();
 
-                $networks = $this->repository->getAllByEnterprise($request->user()->enterprise_id, ['member', 'office']);
+                $networks = $this->repository->getAllByEnterprise($request->user()->enterprise_id, ['member', 'congregation']);
 
                 return response()->json(['networks' => $networks, 'message' => 'Rede cadastrada com sucesso'], 201);
             }
@@ -67,12 +72,16 @@ class NetworkController
     {
         try {
             DB::beginTransaction();
+            if (Auth::user()->enterprise && Auth::user()->enterprise->created_by !== null) {
+                throw new \Exception('Sua organização não tem permissão');
+            }
+
             $network = $this->service->update($request);
 
             if ($network) {
                 DB::commit();
 
-                $networks = $this->repository->getAllByEnterprise($request->user()->enterprise_id, ['member', 'office']);
+                $networks = $this->repository->getAllByEnterprise($request->user()->enterprise_id, ['member', 'congregation']);
 
                 return response()->json(['networks' => $networks, 'message' => 'Rede atualizada com sucesso'], 200);
             }
@@ -91,6 +100,9 @@ class NetworkController
     {
         try {
             DB::beginTransaction();
+            if (Auth::user()->enterprise && Auth::user()->enterprise->created_by !== null) {
+                throw new \Exception('Sua organização não tem permissão');
+            }
 
             $this->rule->delete($id);
             $network = $this->repository->delete($id);
@@ -98,7 +110,7 @@ class NetworkController
             if ($network) {
                 DB::commit();
 
-                $networks = $this->repository->getAllByEnterprise($request->user()->enterprise_id, ['member', 'office']);
+                $networks = $this->repository->getAllByEnterprise($request->user()->enterprise_id, ['member', 'congregation']);
 
                 return response()->json(['networks' => $networks, 'message' => 'Rede excluída com sucesso'], 200);
             }
